ticketapi/requirements.txt
djangorestframework==3.12.2
django==3.1.5
psycopg2==2.7.4
pytz==2018.3
django-cors-headers==3.5.0
pika==1.1.0